Addressed more warnings

+++ xtext/org.icyphy.linguafranca/src/lib/core/reactor_threaded.c
@@ -719,7 +719,7 @@ bool _lf_is_blocked_by_executing_reaction(reaction_t* reaction) {
     if (reaction == NULL) {
         return false;
     }
-    for (int i = 1; i < executing_q->size; i++) {
+    for (size_t i = 1; i < executing_q->size; i++) {
         reaction_t* running = (reaction_t*) executing_q->d[i];
         if (LEVEL(running->index) < LEVEL(reaction->index)
                 && OVERLAPPING(reaction->chain_id, running->chain_id)) {
@@ -1086,7 +1086,7 @@ void start_threads() {
     LOG_PRINT("Starting %d worker threads.", _lf_number_of_threads);
     __thread_ids = (pthread_t*)malloc(_lf_number_of_threads * sizeof(pthread_t));
     number_of_idle_threads = _lf_number_of_threads;
-    for (int i = 0; i < _lf_number_of_threads; i++) {
+    for (unsigned int i = 0; i < _lf_number_of_threads; i++) {
         pthread_create(&__thread_ids[i], NULL, worker, NULL);
     }
 }
